Reapply "Use NotLocalizedString on DBP error screen copy" (#2731)
This reverts commit 71e3c0cd911dcc6240fe1fca54bf560a139ed5d2.
